Move gui transforms onto the 2D matrix stack
GuiGraphics.pose() is a Matrix3x2fStack now: pushMatrix/popMatrix rather
than pushPose/popPose, float arguments, and no z. The depth argument the
screens passed carried no meaning any more since GuiGraphics sorts
render order itself, so it is dropped rather than folded in.

The three sites whose z argument held a computed expression are dropped
by hand; the rest are mechanical.

Error sites 2541 -> 2528.

+++ src/main/java/com/simibubi/create/compat/jei/ItemIcon.java
@@ -37,7 +37,7 @@ public class ItemIcon implements IDrawable {
 		}
 
 		matrixStack.pushMatrix();
-		matrixStack.translate(xOffset + 1, yOffset + 1);
+		matrixStack.translate((float) (xOffset + 1), (float) (yOffset + 1));
 
 		GuiGameElement.of(stack)
 			.submit(graphics);
