+++ src/main/java/com/simibubi/create/compat/jei/category/animations/AnimatedDeployer.java
@@ -18,7 +18,7 @@ public class AnimatedDeployer extends AnimatedKinetics {
 	public void draw(GuiGraphicsExtractor graphics, int xOffset, int yOffset) {
 		Matrix3x2fStack matrixStack = graphics.pose();
 		matrixStack.pushMatrix();
-		matrixStack.translate(xOffset, yOffset);
+		matrixStack.translate((float) (xOffset), (float) (yOffset));
 		// The isometric tilt needs a 3D rotation on what is now a 2D GUI stack.
 		// catnip has not restored that yet, so these draw untilted for now.
 		int scale = 20;
@@ -39,7 +39,7 @@ public class AnimatedDeployer extends AnimatedKinetics {
 
 		matrixStack.pushMatrix();
 
-		matrixStack.translate(0, offset * 17);
+		matrixStack.translate((float) (0), (float) (offset * 17));
 		blockElement(AllPartialModels.DEPLOYER_POLE)
 			.rotateBlock(90, 0, 0)
 			.scale(scale)
