+++ src/main/java/com/simibubi/create/content/schematics/client/SchematicHotbarSlotOverlay.java
@@ -17,7 +17,7 @@ public class SchematicHotbarSlotOverlay  {
 		int y = mainWindow.getGuiScaledHeight() - 19;
 		Matrix3x2fStack ms = graphics.pose();
 		ms.pushMatrix();
-		ms.translate(0, 0);
+		ms.translate((float) (0), (float) (0));
 		AllGuiTextures.SCHEMATIC_SLOT.render(graphics, x + 20 * slot, y);
 		ms.popMatrix();
 	}
